What Datasets are available from covid19R

Returns information about the datasets available from covid19R

get_covid19_data_info(
  location_types_filter = NULL,
  spatial_extent_filter = NULL,
  data_types_filter = NULL
)

Arguments

location_types_filter
A vector of location types that you would like to include. Other location types will be removed. Defaults to NULL, and no filtering is done. See get_covid19_controlled_vocab for how to see possible values.

spatial_extent_filter
A vector of spatial extents that you would like to include. Other extents will be removed. Defaults to NULL, and no filtering is done. See get_covid19_controlled_vocab for how to see possible values.

data_types_filter
A vector of data types that you would like to include. Other data types will be removed. Defaults to NULL, and no filtering is done. See get_covid19_controlled_vocab for how to see possible values.

Value

A tibble listing what datasets are available, and additional information about them.

Examples

get_covid19_data_info()

#> # A tibble: 14 x 14
#>    data_set_name package_name function_to_get… data_details data_url license_url
#>    <chr>         <chr>        <chr>            <chr>        <chr>    <chr>      
#>  1 covid19nytim… covid19nyti… refresh_covid19… "Open Sourc… https:/… https://gi…
#>  2 covid19nytim… covid19nyti… refresh_covid19… "Open Sourc… https:/… https://gi…
#>  3 covid19france covid19fran… refresh_covid19… "Open Sourc… https:/… https://gi…
#>  4 CanadaC19_ca… CanadaC19    refresh_CanadaC… "Open Sourc… https:/… https://gi…
#>  5 covid19us     covid19us    refresh_covid19… "Open Sourc… https:/… https://gi…
#>  6 VirginiaC19   VirginiaC19  refresh_Virgini… "Open Sourc… https:/… https://gi…
#>  7 covid19tunis… covid19tuni… refresh_covid19… "Open Sourc… https:/… https://gi…
#>  8 covid19mobil… covid19mobi… refresh_covid19… "Data refle… https:/… https://ww…
#>  9 covid19mobil… covid19mobi… refresh_covid19… "Data refle… https:/… https://ww…
#> 10 covid19mobil… covid19mobi… refresh_covid19… "Data refle… https:/… https://ww…
#> 11 covid19mobil… covid19mobi… refresh_covid19… "Changes fo… https:/… https://ww…
#> 12 covid19mobil… covid19mobi… refresh_covid19… "Changes fo… https:/… https://ww…
#> 13 covid19mobil… covid19mobi… refresh_covid19… "Changes fo… https:/… https://ww…
#> 14 coronavirus_… coronavirus  refresh_coronav… "The 2019 N… https:/… https://gi…
#> # … with 8 more variables: data_types <chr>, location_types <chr>,
#> #   spatial_extent <chr>, has_geospatial_info <lgl>, TRUE. <lgl>,
#> #   get_info_passing <lgl>, refresh_status <chr>, last_refresh_update <dttm>