Get Information About the Elements of the covid19R controlled vocabularies

Many columns of the tidy format for covid19R data use a controlled vocabulary to define standardized types. While you can read more about these in our documentation, this pulls the controlled vocabulary and description for a given vocabulary type.

get_covid19_controlled_vocab(vocab)

Arguments

vocab
name of a column from a dataset pulled from get_covid19_dataset or a data type listed in get_covid19_data_info

Value

a tibble with information about approved controlled vocabulary terms and their description

Examples

get_covid19_controlled_vocab("data_type")

#> # A tibble: 11 x 2
#>    data_type      description                                                   
#>    <chr>          <chr>                                                         
#>  1 cases_new      new confirmed Covid-19 cases during on the current date       
#>  2 cases_total    cumulative confirmed Covid-19 cases as of the current date    
#>  3 recovered_tot… cumulative number of patients recovered as of the current date
#>  4 recovered_new  new number of patients recovered on the current date          
#>  5 deaths_new     new deaths on the current date                                
#>  6 deaths_total   cumulative deaths due to Covid-19 as of the current date      
#>  7 tested_total   cumulative number of tests performed as of the date           
#>  8 hosp_new       new hospitalizations on the current date                      
#>  9 hosp_current   current number of hospitalized patients as of the current date
#> 10 icu_current    number of hospitalized patients in ICUs as of the current date
#> 11 vent_current   number of hospitalized patients requiring ventilation as of t…